# archsmartadm.indiginfoundation.com/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "ArchSmart API Documentation",
        "description": "Complete API documentation for ArchSmart Real Estate Platform. \r\n * \r\n * **Authentication:**\r\n * 1. First, login using the `/login` endpoint with your credentials\r\n * 2. Copy the token from the response\r\n * 3. Click the 'Authorize' button (🔓) at the top right\r\n * 4. Paste your token in the value field (just the token, without 'Bearer')\r\n * 5. Click 'Authorize' and then 'Close'\r\n * 6. Now you can test all protected endpoints\r\n * \r\n * **Note:** All API requests should use `Accept: application/json` header.",
        "contact": {
            "name": "ArchSmart Support",
            "email": "admin@archsmart.com"
        },
        "license": {
            "name": "Proprietary",
            "url": "https://archsmart.indiginfoundation.com/license"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "https://archsmartadm.indiginfoundation.com/api",
            "description": "Production API Server"
        },
        {
            "url": "http://localhost:8000/api",
            "description": "Local Development Server"
        },
        {
            "url": "/api",
            "description": "Current Domain API Server"
        }
    ],
    "paths": {
...
        "/drafts": {
            "post": {
                "tags": [
                    "Drafts"
                ],
                "summary": "Save blog post draft",
                "description": "Create or update a blog post draft for autosave functionality",
                "operationId": "6e7549d3f56de79d146dbed4e097e7d9",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "id": {
                                        "description": "Draft ID to update (optional)",
                                        "type": "integer"
                                    },
                                    "title": {
                                        "type": "string"
                                    },
                                    "subtitle": {
                                        "type": "string"
                                    },
                                    "excerpt": {
                                        "type": "string"
                                    },
                                    "content": {
                                        "type": "string"
                                    },
                                    "category": {
                                        "type": "string"
                                    },
                                    "tags": {
                                        "type": "array",
                                        "items": {
                                            "type": "string"
                                        }
                                    },
                                    "meta_title": {
                                        "type": "string"
                                    },
                                    "meta_description": {
                                        "type": "string"
                                    },
                                    "canonical_url": {
                                        "type": "string"
                                    },
                                    "is_featured": {
                                        "type": "boolean"
                                    },
                                    "status": {
                                        "type": "string"
                                    },
                                    "published_at": {
                                        "type": "string",
                                        "format": "date-time"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Draft saved",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "integer"
                                        },
                                        "title": {
                                            "type": "string"
                                        },
                                        "user_id": {
                                            "type": "integer"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/drafts/{id}": {
            "get": {
                "tags": [
                    "Drafts"
                ],
                "summary": "Get draft by ID",
                "description": "Retrieve a saved blog post draft",
                "operationId": "2c1b5ec5b8094ed16a2f99d6c2e95833",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Draft ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Draft details",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "integer"
                                        },
                                        "title": {
                                            "type": "string"
                                        },
                                        "content": {
                                            "type": "string"
                                        },
                                        "user_id": {
                                            "type": "integer"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Draft not found"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
    "components": {
...
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "description": "Enter your Bearer token (the token will be automatically prefixed with 'Bearer ')",
                "bearerFormat": "Token",
                "scheme": "bearer"
            }
        }
    },
...
}